Extract transaction-status-client-types (#2870)
* extract ttransaction-status-client-types and transform associated methods to free funcs: UiInstruction::parse -> parse_ui_instruction; UiInnerInstructions::parse => parse_ui_inner_instructions; UiTransactionStatusMeta::parse -> parse_ui_transaction_status_meta

* replace transaction_status with transaction_status_client_types in rpc_client_api

* update description

* also rip out types used by rpc-client

* unused imports

* fix feature activation after workspace change

* move tests

* sort deps
