+++ rpc-client-api/Cargo.toml
@@ -23,7 +23,7 @@ serde_json = { workspace = true }
 solana-account-decoder = { workspace = true }
 solana-inline-spl = { workspace = true }
 solana-sdk = { workspace = true }
-solana-transaction-status = { workspace = true }
+solana-transaction-status-client-types = { workspace = true }
 solana-version = { workspace = true }
 thiserror = { workspace = true }
