+++ rpc-client-api/src/custom_error.rs
@@ -3,7 +3,7 @@ use {
     crate::response::RpcSimulateTransactionResult,
     jsonrpc_core::{Error, ErrorCode},
     solana_sdk::clock::Slot,
-    solana_transaction_status::EncodeError,
+    solana_transaction_status_client_types::EncodeError,
     thiserror::Error,
 };
