+++ transaction-status-client-types/src/option_serializer.rs
